requirements-dev.txt
# Base requirements
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
coverage>=7.3.0

# Code Quality & Linting
flake8>=6.1.0
pylint>=3.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Type stubs
types-PyYAML
pandas-stubs

# Pre-commit hooks
pre-commit>=3.3.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Data science libraries for analytics
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Kedro plugins
kedro-datasets[pandas]>=1.8.0
kedro-viz>=6.7.0
